requirements.txt
anthropic>=0.39.0
chromadb>=0.4.22
streamlit>=1.29.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
tabulate>=0.9.0
reportlab>=4.4.0
